refactor(cli): remove dead/superseded commands watch/daemon/recover/experimental (RFC 7c7859d1 W0)
W0 cleanup from the dogfood homoiconic-command-surface RFC. Removes four
commands that are dead, unwired, or superseded — none have CI/launchd/runtime
consumers (verified by grep across all packages + ci.yml):

- watch        — dead NDJSON file-watcher; FileSystemWatcher had no other
                 consumer, so the whole stack is removed.
- daemon       — whole stack unwired in prod (DaemonClient had no src consumer;
                 ValidatorDaemon only consumed by the daemon command + barrel
                 export). socket-hook degrades to in-process fallback.
- recover      — tmux orphan-recovery, unused (OrphanRecoveryService had no
                 other consumer).
- experimental — rest-push PoC superseded by 'exosync push'. RestPushService
                 is KEPT (used by exosync sync/parity/quarantine + core sync).

Surface: 22 -> 18 commands.

Refactor: createProgram extracted to src/program.ts (side-effect-free, only
src/index.ts calls .parse()) so the real command registry is unit-testable.

Revert-verify (Executable Specification): tests/unit/commands/w0-removed-commands.test.ts
asserts the four commands are absent from the REAL createProgram surface —
empirically RED with them registered, GREEN after removal.

Docs: README command table + sections removed; exosync.md stale 'experimental
rest-push' token-resolution reference updated (file had pre-existing prettier
drift -> whole-file normalised, renders identically).

RFC: vault-exodev/inbox/7c7859d1; WBS node 65751335.
